# lib/api-zod/src/generated/types/announcement.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * HR Management API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Announcement {
  id: number;
  title: string;
  content: string;
  authorId: number;
  /** @nullable */
  authorName?: string | null;
  createdAt: string;
  /** @nullable */
  imageUrl?: string | null;
}