# lib/api-zod/src/generated/types/requestStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * HR Management API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RequestStats {
  total: number;
  pending: number;
  approved: number;
  denied: number;
}