# lib/api-zod/src/generated/types/shiftInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * HR Management API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ShiftInput {
  userId: number;
  date: string;
  startTime: string;
  endTime: string;
  notes?: string;
}